Crate layout_2020

Re-exports

pub use flow::BoxTree;

Modules

cell 🔒

context

display_list

dom

dom_traversal 🔒

flexbox 🔒

flow

Flow layout, also known as block-and-inline layout.

formatting_contexts 🔒

fragment_tree 🔒

geom

layout_debug

Supports writing a trace file created during each layout scope that can be viewed by an external tool to make layout debugging easier.

lists 🔒

positioned 🔒

query

Utilities for querying the layout, as needed by layout.

replaced 🔒

sizing 🔒

https://drafts.csswg.org/css-sizing/

style_ext 🔒

table

HTML Tables (╯°□°)╯︵ ┻━┻

traversal

Macros

layout_debug_scope

Structs

ContainingBlock

DefiniteContainingBlock 🔒

FragmentTree