Module layout_2020::display_list

Modules

background 🔒

conversions 🔒

gradient 🔒

stacking_context 🔒

Structs

BuilderForBoxFragment 🔒

DebugPrintItem

Refers to one of the child contents or stacking contexts of a StackingContext.

DisplayList

Where the information that’s used to build display lists is stored. This includes both a wr::DisplayListBuilder for building up WebRender-specific display list information and a CompositorDisplayListInfo used to store information used by the compositor, such as a compositor-side scroll tree.

DisplayListBuilder 🔒

StackingContext

Either a stacking context or a stacking container, per the definitions in https://drafts.csswg.org/css-position-4/#painting-order.

WebRenderImageInfo

Enums

DebugPrintField

Refers to one of the vecs of a StackingContext.

Functions

clip_for_radii 🔒

cursor 🔒

glyphs 🔒

image_rendering 🔒

inner_radii 🔒

Radii for the padding edge or content edge

offset_radii 🔒

rgba 🔒

Type Aliases

HitInfo 🔒

ItemTag 🔒