Trait layout_2020::dom::NodeExt

pub(crate) trait NodeExt<'dom>: 'dom + LayoutNode<'dom> {
    // Required methods
    fn as_image(self) -> Option<(Option<Arc<NetImage>>, PhysicalSize<f64>)>;
    fn as_canvas(self) -> Option<(CanvasInfo, PhysicalSize<f64>)>;
    fn as_iframe(self) -> Option<(PipelineId, BrowsingContextId)>;
    fn as_video(self) -> Option<(ImageKey, PhysicalSize<f64>)>;
    fn style(self, context: &LayoutContext<'_>) -> ServoArc<ComputedValues>;
    fn get_style_and_layout_data(self) -> Option<StyleAndLayoutData<'dom>>;
    fn layout_data_mut(self) -> AtomicRefMut<'dom, DOMLayoutData>;
    fn element_box_slot(&self) -> BoxSlot<'dom>;
    fn pseudo_element_box_slot(
        &self,
        which: WhichPseudoElement
    ) -> BoxSlot<'dom>;
    fn unset_pseudo_element_box(self, which: WhichPseudoElement);
    fn unset_all_boxes(self);
}

Required Methods

fn as_image(self) -> Option<(Option<Arc<NetImage>>, PhysicalSize<f64>)>

Returns the image if it’s loaded, and its size in image pixels adjusted for image_density.

fn as_canvas(self) -> Option<(CanvasInfo, PhysicalSize<f64>)>

fn as_iframe(self) -> Option<(PipelineId, BrowsingContextId)>

fn as_video(self) -> Option<(ImageKey, PhysicalSize<f64>)>

fn style(self, context: &LayoutContext<'_>) -> ServoArc<ComputedValues>

fn get_style_and_layout_data(self) -> Option<StyleAndLayoutData<'dom>>

fn layout_data_mut(self) -> AtomicRefMut<'dom, DOMLayoutData>

fn element_box_slot(&self) -> BoxSlot<'dom>

fn pseudo_element_box_slot(&self, which: WhichPseudoElement) -> BoxSlot<'dom>

fn unset_pseudo_element_box(self, which: WhichPseudoElement)

fn unset_all_boxes(self)

Remove boxes for the element itself, and its :before and :after if any.