Module layout_2020::flexbox

Modules

construct 🔒

geom 🔒

https://drafts.csswg.org/css-flexbox/#box-model

layout 🔒

Structs

CachedBlockSizeContribution 🔒

FlexContainer 🔒

FlexContainerConfig 🔒

A structure to hold the configuration of a flex container for use during layout and preferred width calculation.

FlexItemBox 🔒

Enums

FlexLevelBox 🔒