Module layout_2020::flexbox::layout

Structs

DesiredFlexFractionAndGrowOrShrinkFactor 🔒

FinalFlexLineLayout 🔒

Return type of FlexLine::layout

FlexContext 🔒

Layout parameters and intermediate results about a flex container, grouped to avoid passing around many parameters

FlexItem 🔒

A flex item with some intermediate results

FlexItemBoxInlineContentSizesInfo 🔒

FlexItemLayoutResult 🔒

Return type of FlexItem::layout

InitialFlexLineLayout 🔒

Enums

FlexContent 🔒

Child of a FlexContainer. Can either be absolutely positioned, or not. If not, a placeholder is used and flex content is stored outside of this enum.

Functions

allocate_free_cross_space_for_flex_line 🔒

Align all flex lines per align-content according to https://drafts.csswg.org/css-flexbox/#algo-line-align. Returns the space to add to each line or the space to add after each line.

cross_axis_is_item_block_axis 🔒

do_initial_flex_line_layout 🔒

item_with_auto_cross_size_stretches_to_line_size 🔒

Whether an item with an auto preferred cross size will stretched to fill the cross size of its flex line. https://drafts.csswg.org/css-flexbox/#stretched