Function layout_2020::flow::calculate_inline_content_size_for_block_level_boxes

fn calculate_inline_content_size_for_block_level_boxes(
    boxes: &[ArcRefCell<BlockLevelBox>],
    layout_context: &LayoutContext<'_>,
    containing_block: &IndefiniteContainingBlock<'_>,
) -> InlineContentSizesResult

Finds the min/max-content inline size of the block-level children of a block container. The in-flow boxes will stack vertically, so we only need to consider the maximum size. But floats can flow horizontally depending on ‘clear’, so we may need to sum their sizes. CSS 2 does not define the exact algorithm, this logic is based on the behavior observed on Gecko and Blink.