Module layout_2020::flow::inline::text_run

Structs

TextRun 🔒

https://www.w3.org/TR/css-display-3/#css-text-run

TextRunSegment 🔒

TwoCharsAtATimeIterator 🔒

Enums

SegmentStartSoftWrapPolicy 🔒

Constants

XI_LINE_BREAKING_CLASS_CM 🔒

XI_LINE_BREAKING_CLASS_GL 🔒

XI_LINE_BREAKING_CLASS_WJ 🔒

XI_LINE_BREAKING_CLASS_ZW 🔒

XI_LINE_BREAKING_CLASS_ZWJ 🔒

Functions

add_or_get_font 🔒

char_does_not_change_font 🔒

Whether or not this character should be able to change the font during segmentation. Certain character are not rendered at all, so it doesn’t matter what font we use to render them. They should just be added to the current segment.

get_font_for_first_font_for_style 🔒