Module layout_2020::flow::text_run

Structs

TextRun 🔒

https://www.w3.org/TR/css-display-3/#css-text-run

TextRunSegment 🔒

TextTransformation

This is an interator that consumes a char iterator and produces character transformed by the given CSS text-transform value. It currently does not support text-transform: capitalize because Unicode segmentation libraries do not support streaming input one character at a time.

WhitespaceCollapse

Enums

PendingCaseConversionResult 🔒

SegmentStartSoftWrapPolicy 🔒

Constants

XI_LINE_BREAKING_CLASS_CM 🔒

XI_LINE_BREAKING_CLASS_GL 🔒

XI_LINE_BREAKING_CLASS_WJ 🔒

XI_LINE_BREAKING_CLASS_ZW 🔒

XI_LINE_BREAKING_CLASS_ZWJ 🔒

Functions

add_or_get_font 🔒

capitalize_string 🔒

Given a string and whether the start of the string represents a word boundary, create a copy of the string with letters after word boundaries capitalized.

char_does_not_change_font 🔒

Whether or not this character should be able to change the font during segmentation. Certain character are not rendered at all, so it doesn’t matter what font we use to render them. They should just be added to the current segment.

char_prevents_soft_wrap_opportunity_when_before_or_after_atomic 🔒

Whether or not this character will rpevent a soft wrap opportunity when it comes before or after an atomic inline element.

get_font_for_first_font_for_style 🔒

preserve_segment_break 🔒