Module layout_2020::query

Utilities for querying the layout, as needed by layout.

Structs

RenderedTextCollectionState 🔒

Enums

InnerOrOuterTextItem 🔒

Functions

get_the_text_steps

https://html.spec.whatwg.org/multipage/#get-the-text-steps

is_eligible_parent 🔒

Returns whether or not the element with the given style and body element determination is eligible to be a parent element for offset* queries.

process_content_box_request

process_content_boxes_request

process_node_geometry_request

process_node_scroll_area_request

https://drafts.csswg.org/cssom-view/#scrolling-area

process_offset_parent_query

process_offset_parent_query_inner 🔒

process_resolved_font_style_query

process_resolved_style_request

Return the resolved value of property for a given (pseudo)element. https://drafts.csswg.org/cssom/#resolved-value

process_resolved_style_request_for_unstyled_node

process_text_index_request

rendered_text_collection_steps 🔒

https://html.spec.whatwg.org/multipage/#rendered-text-collection-steps

resolved_size_should_be_used_value 🔒

shorthand_to_css_string 🔒