Module layout_2020::replaced

Structs

CanvasInfo 🔒

IFrameInfo 🔒

IntrinsicSizes 🔒

Raster images always have an intrinsic width and height, with 1 image pixel = 1px. The intrinsic ratio should be based on dividing those. See https://github.com/w3c/csswg-drafts/issues/4572 for the case where either is zero. PNG specifically disallows this but I (SimonSapin) am not sure about other formats.Form controls have both intrinsic width and height but no intrinsic ratio. See https://github.com/w3c/csswg-drafts/issues/1044 and https://drafts.csswg.org/css-images/#intrinsic-dimensions “In general, […]”For SVG, see https://svgwg.org/svg2-draft/coords.html#SizingSVGInCSS and again https://github.com/w3c/csswg-drafts/issues/4572.IFrames do not have intrinsic width and height or intrinsic ratio according to https://drafts.csswg.org/css-images/#intrinsic-dimensions.

ReplacedContent 🔒

VideoInfo 🔒

Enums

CanvasSource 🔒

ReplacedContentKind 🔒