Module layout_2020::table::layout

Structs

CellLayout 🔒

A result of a final or speculative layout of a single cell in the table. Note that this is only done for slots that are not covered by spans or empty.

CellOrTrackMeasure 🔒

CollapsedBorder 🔒

A calculated collapsed border.

CollapsedBorders 🔒

The calculated collapsed borders.

ColumnLayout 🔒

Information stored during the layout of columns.

RowFragmentLayout 🔒

RowGroupFragmentLayout 🔒

RowLayout 🔒

Information stored during the layout of rows.

RowspanToDistribute 🔒

TableAndTrackDimensions 🔒

TableLayout 🔒

A helper struct that performs the layout of the box tree version of a table into the fragment tree version. This implements https://drafts.csswg.org/css-tables/#table-layout-algorithm

Functions

get_outer_sizes_for_measurement 🔒

get_size_percentage_contribution 🔒