Type Alias layout_2020::table::TableSize

pub type TableSize = Size2D<usize, UnknownUnit>;

Aliased Type

struct TableSize {
    pub width: usize,
    pub height: usize,
}

Fields
width: usize

The extent of the element in the U units along the x axis (usually horizontal).

height: usize

The extent of the element in the U units along the y axis (usually vertical).