Module layout_2020::flow

Flow layout, also known as block-and-inline layout.

Modules

construct 🔒

float

Float layout.

inline

Inline Formatting Context Layout

root 🔒

Structs

BlockFormattingContext 🔒

BoxTree

CanvasBackground

https://drafts.csswg.org/css-backgrounds/#root-background

CollapsibleWithParentStartMargin 🔒

ContainingBlockPaddingAndBorder 🔒

FlowLayout 🔒

OutsideMarker 🔒

The contentes of a BlockContainer created to render a list marker for a list that has list-style-position: outside.

PlacementState 🔒

State that we maintain when placing blocks.

ResolvedMargins 🔒

Enums

BlockContainer 🔒

BlockLevelBox 🔒

Functions

block_size_is_zero_or_intrinsic 🔒

calculate_inline_content_size_for_block_level_boxes 🔒

Finds the min/max-content inline size of the block-level children of a block container. The in-flow boxes will stack vertically, so we only need to consider the maximum size. But floats can flow horizontally depending on ‘clear’, so we may need to sum their sizes. CSS 2 does not define the exact algorithm, this logic is based on the behavior observed on Gecko and Blink.

justify_self_alignment 🔒

This is supposed to handle ‘justify-self’, but no browser supports it on block boxes. Instead, <center> and <div align> are implemented via internal ‘text-align’ values. The provided free space should already take margins into account. In particular, it should be zero if there is an auto margin. https://drafts.csswg.org/css-align/#justify-block

layout_block_level_children 🔒

layout_block_level_children_in_parallel 🔒

layout_block_level_children_sequentially 🔒

layout_in_flow_non_replaced_block_level_same_formatting_context 🔒

Lay out a normal flow non-replaced block that does not establish a new formatting context.

layout_in_flow_replaced_block_level 🔒

https://drafts.csswg.org/css2/visudet.html#block-replaced-width https://drafts.csswg.org/css2/visudet.html#inline-replaced-width https://drafts.csswg.org/css2/visudet.html#inline-replaced-height

solve_block_margins_for_in_flow_block_level 🔒

Resolves ‘auto’ margins of an in-flow block-level box in the block axis. https://drafts.csswg.org/css2/#normal-block https://drafts.csswg.org/css2/#block-root-margin

solve_clearance_and_inline_margins_avoiding_floats 🔒

A block-level element that establishes an independent formatting context (or is replaced) must not overlap floats. This can be achieved by adding clearance (to adjust the position in the block axis) and/or modifying the margins in the inline axis. This function takes care of calculating them.

solve_containing_block_padding_and_border_for_in_flow_box 🔒

Given the style for an in-flow box and its containing block, determine the containing block for its children. Note that in the presence of floats, this shouldn’t be used for a block-level box that establishes an independent formatting context (or is replaced), since the inline size could then be incorrect.

solve_inline_margins_avoiding_floats 🔒

Resolves ‘auto’ margins of an in-flow block-level box in the inline axis similarly to |solve_inline_margins_for_in_flow_block_level|. However, they align within the provided rect (instead of the containing block), to avoid overlapping floats. In addition to the used margins, it also returns the effective margin-inline-start (see ContainingBlockPaddingAndBorder). It may differ from the used inline-start margin if the computed value wasn’t ‘auto’ and there are floats to avoid or the box is justified. See https://github.com/w3c/csswg-drafts/issues/9174

solve_inline_margins_for_in_flow_block_level 🔒

Resolves ‘auto’ margins of an in-flow block-level box in the inline axis, distributing the free space in the containing block.

solve_margins 🔒

Given the containing block and size of an in-flow box, determine the margins. Note that in the presence of floats, this shouldn’t be used for a block-level box that establishes an independent formatting context (or is replaced), since the margins could then be incorrect.